Common project is internal Exposed only to other SurrealDB project via InternalsVisibleTo in AssemblyInfo.cs
