+++ src/Common/AssemblyInfo.cs
@@ -1,0 +1,12 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("SurrealDB.Abstractions")]
+[assembly: InternalsVisibleTo("SurrealDB.Configuration")]
+[assembly: InternalsVisibleTo("SurrealDB.Driver.Rest")]
+[assembly: InternalsVisibleTo("SurrealDB.Driver.Rpc")]
+[assembly: InternalsVisibleTo("SurrealDB.Models")]
+[assembly: InternalsVisibleTo("SurrealDB.Json")]
+[assembly: InternalsVisibleTo("SurrealDB.Ws")]
+[assembly: InternalsVisibleTo("SurrealDB.Core.Tests")]
+[assembly: InternalsVisibleTo("SurrealDB.Driver.Tests")]
+[assembly: InternalsVisibleTo("SurrealDB.Shared.Tests")]
