+++ src/Common/TimeZoneHelper.cs
@@ -1,44 +1,6 @@
-namespace SurrealDB.Json;
+namespace SurrealDB.Common;
 
-public static class TimeZoneHelper {
-    private static Dictionary<TimeSpan, TimeZoneInfo>? s_cacheTz;
-    private static readonly object s_cacheTzLock = new();
-
-    private static IReadOnlyDictionary<TimeSpan, TimeZoneInfo> GetTimeZones() {
-        Dictionary<TimeSpan, TimeZoneInfo> cache;
-        lock (s_cacheTzLock) {
-            cache = EnsureTzCacheUnchecked();
-        }
-
-        return cache;
-    }
-
-    private static Dictionary<TimeSpan, TimeZoneInfo> EnsureTzCacheUnchecked() {
-        Dictionary<TimeSpan, TimeZoneInfo>? cache = s_cacheTz;
-        if (cache is null || cache.Count <= 0) {
-            cache = TimeZoneInfo.GetSystemTimeZones().ToDictionary(static tz => tz.BaseUtcOffset);
-            s_cacheTz = cache;
-        }
-
-        return cache;
-    }
-
-    private static TimeZoneInfo AddTimeZone(string name, in TimeSpan off) {
-        lock (s_cacheTzLock) {
-            Dictionary<TimeSpan, TimeZoneInfo> cache = EnsureTzCacheUnchecked();
-            if (cache.TryGetValue(off, out TimeZoneInfo? tz)) {
-                return tz;
-            }
-            tz = TimeZoneInfo.CreateCustomTimeZone(name, off, null, null, null, null, false);
-            cache[off] = tz;
-            return tz;
-        }
-    }
-
-    public static TimeZoneInfo FromOffset(in TimeSpan offset) {
-        return GetTimeZones().TryGetValue(offset, out TimeZoneInfo? tz) ? tz : AddTimeZone(offset.ToString(), in offset);
-    }
-
+internal static class TimeZoneHelper {
     /// <summary>
     /// Converts the <see cref="DateTime"/> to a <see cref="DateTimeOffset"/> by adding a offset.
     /// </summary>
