+++ tests/Driver.Tests/Queries/GeneralQueryTests.cs
@@ -88,7 +88,7 @@ GROUP BY country;";
         };
         var expectedResult = new MathResultDocument { result = expectedObject.f1 + expectedObject.f2 };
 
-        Thing thing = Thing.From("object", RngHelper.Shared.Next().ToString());
+        Thing thing = Thing.From("object", ThreadRng.Shared.Next().ToString());
         await Database.Create(thing, expectedObject);
 
         string sql = "SELECT (f1 + f2) as result FROM $record";
@@ -112,7 +112,7 @@ GROUP BY country;";
         };
         var expectedResult = new MathResultDocument { result = expectedObject.f1 + expectedObject.f2 };
 
-        Thing thing = Thing.From("object", RngHelper.Shared.Next().ToString());
+        Thing thing = Thing.From("object", ThreadRng.Shared.Next().ToString());
         await Database.Create(thing, expectedObject);
 
         string sql = "SELECT (f1 + f2) as result FROM $record";
@@ -137,7 +137,7 @@ GROUP BY country;";
         };
         var expectedResult = new MathResultDocument { result = expectedObject.f1 + expectedObject.f2 };
 
-        Thing thing = Thing.From("object", RngHelper.Shared.Next().ToString());
+        Thing thing = Thing.From("object", ThreadRng.Shared.Next().ToString());
         await Database.Create(thing, expectedObject);
 
         string sql = "SELECT (f1 + f2) as result FROM $record";
@@ -162,7 +162,7 @@ GROUP BY country;";
         };
         var expectedResult = new MathResultDocument { result = expectedObject.f1 - expectedObject.f2 };
 
-        Thing thing = Thing.From("object", RngHelper.Shared.Next().ToString());
+        Thing thing = Thing.From("object", ThreadRng.Shared.Next().ToString());
         await Database.Create(thing, expectedObject);
 
         string sql = "SELECT (f1 - f2) as result FROM $record";
